University project

+++ dmitrii/src/main/java/entities/university/SubjectWithMark.java
@@ -4,13 +4,13 @@ package entities.university;
  * Created by mit_OK! on 04.04.2015.
  */
 public class SubjectWithMark extends Subject {
-    private int mark;
+    private Double mark;
 
-    public int getMark() {
+    public Double getMark() {
         return mark;
     }
 
-    public void setMark(int mark) {
+    public void setMark(Double mark) {
         this.mark = mark;
     }
